+++ v2/metrics/processor/types_test.go
@@ -24,10 +24,6 @@ func (f *fakeRegistry) Unregister(c prometheus.Collector) bool {
 	panic("implement me")
 }
 
-func getReceiverName() string {
-	return "testReceiverName"
-}
-
 func TestRegistry_Init(t *testing.T) {
 	g := NewWithT(t)
 	r := NewRegistry()
@@ -35,7 +31,7 @@ func TestRegistry_Init(t *testing.T) {
 	g.Expect(func() { r.Init(prometheus.NewRegistry()) }).ToNot(Panic())
 	g.Expect(func() { r.Init(fRegistry) }).ToNot(Panic())
 	g.Expect(fRegistry.collectors).To(HaveLen(5))
-	Metric.IncMessageReceived(getReceiverName(), 10)
+	Metric.IncMessageReceived("testReceiverName", 10)
 }
 
 func TestNewInformerDefault(t *testing.T) {
