--- v2/receiver.go
@@ -1,33 +1,0 @@
-package shuttle
-
-import (
-	"context"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
-)
-
-type Receiver interface {
-	ReceiveMessages(ctx context.Context, maxMessages int, options *azservicebus.ReceiveMessagesOptions) ([]*azservicebus.ReceivedMessage, error)
-	MessageSettler
-}
-
-// MessageSettler is passed to the handlers. it exposes the message settling functionality from the receiver needed within the handler.
-type MessageSettler interface {
-	AbandonMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.AbandonMessageOptions) error
-	CompleteMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.CompleteMessageOptions) error
-	DeadLetterMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.DeadLetterOptions) error
-	DeferMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.DeferMessageOptions) error
-	RenewMessageLock(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.RenewMessageLockOptions) error
-}
-
-type ReceiverImpl struct { // shuttle.Receiver is already an exported interface
-	name       string
-	sbReceiver Receiver
-}
-
-func NewReceiverImpl(name string, sbReceiver Receiver) *ReceiverImpl {
-	return &ReceiverImpl{
-		name:       name,
-		sbReceiver: sbReceiver,
-	}
-}
